Extract EntityProvider into its own class file
